feat: use new error-friendly request rpc client in backend
